Adapt previous code to new model
We use the same method as in the previous code written for the ancient model, and recompute the new tradeoffs (changing some model parameters and thus the borders) to deduce the exclusive consumption zones of each good. We finally derive consumption shares, quantities as well as total environmental impacts (focusing on exclusive consumption only in the first place).